chore: sets all varibles to float annotation

+++ 0x00-python_variable_annotations/0-add.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python3
-def add(a: int, b: float) -> float:
+def add(a: float, b: float) -> float:
     """This function two numbers of different data type"""
     return a + b
